rename core components
